requirements.txt
streamlit>=1.36
langchain>=0.2.12
langchain-community>=0.2.12
langchain-google-genai>=2.1.0
google-generativeai>=0.7.2
chromadb>=0.5.3
pypdf>=4.2.0
python-dotenv>=1.0.1
pandas>=2.2.2
sentence-transformers>=3.0.1
torch>=2.3.1
